refactor: no logic change - move get_learners_search_queryset
