+++ futurex_openedx_extensions/dashboard/details/learners.py
@@ -12,7 +12,11 @@ from lms.djangoapps.certificates.models import GeneratedCertificate
 from lms.djangoapps.courseware.models import StudentModule
 from lms.djangoapps.grades.models import PersistentCourseGrade
 
-from futurex_openedx_extensions.helpers.querysets import check_staff_exist_queryset, get_base_queryset_courses
+from futurex_openedx_extensions.helpers.querysets import (
+    check_staff_exist_queryset,
+    get_base_queryset_courses,
+    get_learners_search_queryset,
+)
 from futurex_openedx_extensions.helpers.tenants import get_tenants_sites
 
 
@@ -89,46 +93,6 @@ def get_certificates_count_for_learner_queryset(
         ),
         distinct=True
     )
-
-
-def get_learners_search_queryset(
-    search_text: str | None = None,
-    superuser_filter: bool | None = False,
-    staff_filter: bool | None = False,
-    active_filter: bool | None = True
-) -> QuerySet:
-    """
-    Get the learners queryset for the given search text.
-
-    :param search_text: Search text to filter the learners by
-    :type search_text: str | None
-    :param superuser_filter: Value to filter superusers. None means no filter
-    :type superuser_filter: bool | None
-    :param staff_filter: Value to filter staff users. None means no filter
-    :type staff_filter: bool | None
-    :param active_filter: Value to filter active users. None means no filter
-    :type active_filter: bool | None
-    :return: QuerySet of learners
-    :rtype: QuerySet
-    """
-    queryset = get_user_model().objects.all()
-
-    if superuser_filter is not None:
-        queryset = queryset.filter(is_superuser=superuser_filter)
-    if staff_filter is not None:
-        queryset = queryset.filter(is_staff=staff_filter)
-    if active_filter is not None:
-        queryset = queryset.filter(is_active=active_filter)
-
-    search_text = (search_text or '').strip()
-    if search_text:
-        queryset = queryset.filter(
-            Q(username__icontains=search_text) |
-            Q(email__icontains=search_text) |
-            Q(profile__name__icontains=search_text)
-        )
-
-    return queryset
 
 
 def get_learners_queryset(
